Rename EditorBox to FieldSet & Replace
