+++ packages/ui/src/components/Forms/FieldSet/FieldSet.sass
@@ -1,0 +1,6 @@
+.#{$cui-conf-globalPrefix}fieldSet
+	display: flex
+	flex-direction: column
+
+	&-heading + &-content
+		margin-top: .75em
