+++ packages/ui/src/components/Forms/FieldSet/index.sass
@@ -1,0 +1,1 @@
+@import './FieldSet'
