+++ packages/ui/src/components/Forms/index.sass
@@ -2,3 +2,4 @@
 @import 'Checkbox/index'
 @import 'SlugControl/index'
 @import 'FieldContainer/index'
+@import 'FieldSet/index'
