+++ packages/ui/src/components/Forms/index.ts
@@ -2,6 +2,7 @@ export * from './Button'
 export * from './Checkbox'
 export * from './ErrorList'
 export * from './FieldContainer'
+export * from './FieldSet'
 export * from './RadioGroup'
 export * from './Select'
 export * from './SlugControl'
