--- packages/ui/src/styles/components/editorBox.sass
@@ -1,14 +1,0 @@
-@import '../../components/Box/boxContentMixin'
-
-.#{$cui-conf-globalPrefix}editorBox
-	+boxContentMixin
-	display: flex
-	flex-direction: column
-
-	&-heading + &-content
-		margin-top: .75em
-
-	&.is-active
-		transition: box-shadow .15s
-		box-shadow: 0 0 0 .2em rgba($cui-control-view-default-distinction-color, .3)
-		z-index: 1
