+++ packages/ui/src/styles/components/index.sass
@@ -9,7 +9,6 @@
 @import './dropdown'
 @import './editorBlock'
 @import './editorBlockBoundary'
-@import './editorBox'
 @import './editorCanvas'
 @import './editorHeading'
 @import './editorNonEditable'
